+++ ExampleGame/Player.cs
@@ -1,6 +1,5 @@
 ﻿using SadRogue.Integration;
 using SadRogue.Integration.Keybindings;
-using SadRogue.Primitives;
 
 namespace ExampleGame
 {
@@ -35,7 +34,7 @@ namespace ExampleGame
         }
 
         // If the player is added to a map, update the player FOV when the player moves
-        private void OnPositionChanged(object? sender, SadConsole.ValueChangedEventArgs<Point> e)
+        private void OnPositionChanged(object? sender, ValueChangedEventArgs<Point> e)
         {
             CalculateFOV();
         }